Split onUpdate* function into separate file
