+++ libs/game/game.ts
@@ -232,40 +232,6 @@ namespace game {
         }
     }
 
-    /**
-     * Update the position and velocities of sprites
-     * @param body code to execute
-     */
-    //% group="Gameplay"
-    //% help=game/on-update weight=100 afterOnStart=true
-    //% blockId=gameupdate block="on game update"
-    //% blockAllowMultiple=1
-    export function onUpdate(a: () => void): void {
-        init();
-        if (!a) return;
-        game.eventContext().registerFrameHandler(scene.UPDATE_PRIORITY, a);
-    }
-
-    /**
-     * Run code on an interval of time. This executes before game.onUpdate()
-     * @param body code to execute
-     */
-    //% group="Gameplay"
-    //% help=game/on-update-interval weight=99 afterOnStart=true
-    //% blockId=gameinterval block="on game update every %period=timePicker ms"
-    //% blockAllowMultiple=1
-    export function onUpdateInterval(period: number, a: () => void): void {
-        init();
-        if (!a || period < 0) return;
-        let timer = 0;
-        game.eventContext().registerFrameHandler(scene.UPDATE_INTERVAL_PRIORITY, () => {
-            const time = game.currentScene().millis();
-            if (timer <= time) {
-                timer = time + period;
-                a();
-            }
-        });
-    }
 
     // Indicates whether the fiber needs to be created
     let foreverRunning = false;
@@ -324,15 +290,6 @@ namespace game {
         scene.createRenderable(scene.ON_SHADE_Z, a);
     }
 
-    /**
-     * Returns the time since the game started in milliseconds
-     */
-    //% blockId=arcade_game_runtime block="time since start (ms)"
-    //% group="Gameplay" weight=11
-    //% help=game/runtime
-    export function runtime(): number {
-        return currentScene().millis();
-    }
 
     /**
      * Register a handler that runs whenever a scene is pushed onto the scene
