+++ libs/game/pxt.json
@@ -7,6 +7,7 @@
         "constants.ts",
         "controlleroverrides.ts",
         "controller.ts",
+        "controllerbutton.ts",
         "hitbox.ts",
         "sprites.ts",
         "sprite.ts",
@@ -27,6 +28,7 @@
         "scenes.ts",
         "textDialogs.ts",
         "game.ts",
+        "gameutil.ts",
         "prompt.ts",
         "numberprompt.ts",
         "ask.ts",
